requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4>=4.12.0
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
iniconfig==2.3.0
multidict==6.7.0
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
Pygments==2.19.2
python-dotenv==1.2.1
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
yarl==1.22.0
